refactor: move modules to separated folder
