+++ src/modules/resume-builder/resume-builder.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { GoogleModule } from 'src/google/google.module';
+import { GoogleModule } from 'src/modules/google/google.module';
 import { ResumeBuilderController } from './resume-builder.controller';
 import { ResumeBuilderService } from './resume-builder.service';
