+++ src/modules/resume-builder/resume-builder.service.ts
@@ -11,7 +11,7 @@ import { GoogleDocTemplateFields, GoogleReplaceRequest } from 'src/types/';
 import { Req, Res } from 'src/dtos';
 import { API_RESPONSE_MESSAGES } from 'src/constants/api-response-messages';
 import { mockProfessionalData } from './__mock__/mock-professional-data';
-import { GoogleService } from 'src/google/google.service';
+import { GoogleService } from 'src/modules/google/google.service';
 import { plainToInstance } from 'class-transformer';
 
 @Injectable()
